backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
async-timeout==5.0.1
asyncpg==0.30.0
certifi==2024.12.14
click==8.1.8
coverage==7.6.10
exceptiongroup==1.2.2
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pg8000==1.31.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
redis==5.2.1
scramp==1.4.5
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.37
starlette==0.45.3
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
WTForms==3.1.2
